+++ app/src/main/java/com/example/money/model/SaveCost.kt
@@ -3,7 +3,7 @@ package com.example.money.model
 class SaveCost
 {
     companion object{
-        var Expence: Int  = 0
-        var Income: Int = 0
+        var Expence: Long  = 0
+        var Income: Long = 0
     }
 }
